SDI-720 create mapping server endpoints to clear activities and appointments dev data

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonertonomisupdate/activities/ActivitiesResource.kt
@@ -6,9 +6,13 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
+import uk.gov.justice.digital.hmpps.prisonertonomisupdate.listeners.EventFeatureSwitch
 
 @RestController
-class ActivitiesResource(private val activitiesService: ActivitiesService) {
+class ActivitiesResource(
+  private val activitiesService: ActivitiesService,
+  private val eventFeatureSwitch: EventFeatureSwitch,
+) {
 
   /**
    * For dev environment only - delete all activities and courses, for when activities environment is reset
@@ -18,6 +22,10 @@ class ActivitiesResource(private val activitiesService: ActivitiesService) {
   @DeleteMapping("/activities")
   @ResponseStatus(HttpStatus.NO_CONTENT)
   suspend fun deleteAllActivities() {
-    activitiesService.deleteAllActivities()
+    if (eventFeatureSwitch.isEnabled("DELETEALL")) {
+      activitiesService.deleteAllActivities()
+    } else {
+      throw RuntimeException("Attempt to delete activities in wrong environment")
+    }
   }
 }
